requirements.txt
flask==2.1.3
werkzeug==2.1.2
python-dotenv==0.21.0